feat: add toggle for React
Content scripts execute automatically when the page is loaded.

As such, we can't just use the default React example directly.
https://github.com/wxt-dev/examples/tree/main/examples/react-content-script-ui/entrypoints/content

Instead, we use the `create` function to defer the creation of
the shadow root, allowing us to mount the UI on demand.
We then add listeners and send a message when the user clicks the
extension icon, allowing us to detect whether the current
state is `ON` or `OFF`.