feat: add bookmark messaging
Necessary to use the messaging infrastucture because the content script
will create the IndexedDB in the current tab's origin
(e.g. https://medium.com) rather than the chrome extension's origin.

To be able to write to our own chrome extension IndexedDB, we need
to send the message to and from the background script.