fix: change styling from rem to em
We cannot use rem for shadow root in a content script because tailwind's
rem styling is influenced by the parent site's `html` element.

The `html` element on some sites sets the font size to a percentage,
like 62.5%, which affects the styling of our side panel.

To handle this, we have to set a root font size for our shadow, and then
set all of our sizing relative to that root font size.

This has only been done for the components affected by the side panel,
and a more robust solution may be needed in the future.